move gatherOsMetrics and sendMetrics outisde main component
