+++ .gitignore
@@ -1,3 +1,4 @@
 index.rendered.html
 node_modules
 coverage
+*.log
