+++ examples/index.js
@@ -1,23 +1,11 @@
+/* eslint no-console: "off" */
+
 const express = require('express');
+
 const app = express();
 
-const config = {
-  path: '/',
-  title: 'Express Status',
-  spans: [{
-    interval: 1,
-    retention: 60
-  }, {
-    interval: 5,
-    retention: 60
-  }, {
-    interval: 15,
-    retention: 60
-  }]
-}
-
-app.use(require('../index')(config));
+app.use(require('../index')({ path: '/' }));
 
 app.listen(3000, () => {
-  console.log('Listening on http://0.0.0.0:3000');
+  console.log('listening on http://0.0.0.0:3000');
 });
